requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Model Persistence
joblib>=1.3.0

# NLP
fasttext-wheel>=0.9.2

# Progress Bars
tqdm>=4.66.0
